[minor] moved precision to base_document
