+++ frappe/model/base_document.py
@@ -395,6 +395,36 @@ class BaseDocument(object):
 				frappe.throw(_("Not allowed to change {0} after submission").format(df.label),
 					frappe.UpdateAfterSubmitError)
 
+	def precision(self, fieldname, parentfield=None):
+		"""Returns float precision for a particular field (or get global default).
+
+		:param fieldname: Fieldname for which precision is required.
+		:param parentfield: If fieldname is in child table."""
+		from frappe.model.meta import get_field_precision
+
+		if parentfield and not isinstance(parentfield, basestring):
+			parentfield = parentfield.parentfield
+
+		cache_key = parentfield or "main"
+
+		if not hasattr(self, "_precision"):
+			self._precision = frappe._dict()
+
+		if cache_key not in self._precision:
+			self._precision[cache_key] = frappe._dict()
+
+		if fieldname not in self._precision[cache_key]:
+			self._precision[cache_key][fieldname] = None
+
+			doctype = self.meta.get_field(parentfield).options if parentfield else self.doctype
+			df = frappe.get_meta(doctype).get_field(fieldname)
+
+			if df.fieldtype in ("Currency", "Float", "Percent"):
+				self._precision[cache_key][fieldname] = get_field_precision(df, self)
+
+		return self._precision[cache_key][fieldname]
+
+
 	def get_formatted(self, fieldname, doc=None, currency=None):
 		from frappe.utils.formatters import format_value
 		return format_value(self.get(fieldname), self.meta.get_field(fieldname),
